+++ dejafu/Test/DejaFu/Internal.hs
@@ -357,6 +357,12 @@ efromRight :: HasCallStack => Either a b -> b
 efromRight (Right b) = b
 efromRight _ = withFrozenCallStack $ fatal "fromRight: Left"
 
+-- | 'fromLeft' but with a better error message if it fails.  Use
+-- this only where it shouldn't fail!
+efromLeft :: HasCallStack => Either a b -> a
+efromLeft (Left a) = a
+efromLeft _ = withFrozenCallStack $ fatal "fromLeft: Right"
+
 -- | 'M.adjust' but which errors if the key is not present.  Use this
 -- only where it shouldn't fail!
 eadjust :: (Ord k, Show k, HasCallStack) => (v -> v) -> k -> M.Map k v -> M.Map k v